+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 
 - color variables in BaseSelect icon and text
+- maxlength in BaseInput and BaseTextarea
 
 ## [0.11.0]